refactor(starfish-core): rename scoring_metrics module and column family to misbehavior_*
The runtime type was renamed `ScoringMetricsStore` → `MisbehaviorStore`
and the storage type `StorageScoringMetrics` → `MisbehaviorCounts` in
the earlier commit; finish that rename for two name stragglers:

- Move `scoring_metrics_store.rs` → `misbehavior_store.rs` via
  `git mv`, and update the module path everywhere
  (`crate::scoring_metrics_store` → `crate::misbehavior_store`).
- Rename the RocksDB column family string `"scoring_metrics"` →
  `"misbehavior_counts"`. Safe to change because the column family
  was introduced in PR #10088, which has not merged.
