+++ crates/starfish/core/src/authority_service.rs
@@ -44,7 +44,7 @@ use crate::{
         SerializedBlockBundleParts, SerializedHeaderAndTransactions, SerializedTransactionsV1,
         SerializedTransactionsV2, TransactionFetchMode,
     },
-    scoring_metrics_store::MisbehaviorStore,
+    misbehavior_store::MisbehaviorStore,
     shard_reconstructor::TransactionMessage,
     stake_aggregator::{QuorumThreshold, StakeAggregator},
     storage::Store,
@@ -1684,7 +1684,7 @@ mod tests {
             SerializedBlockBundle, SerializedBlockBundleParts, SerializedHeaderAndTransactions,
             SerializedTransactionsV1, SerializedTransactionsV2, TransactionFetchMode,
         },
-        scoring_metrics_store::MisbehaviorStore,
+        misbehavior_store::MisbehaviorStore,
         storage::{Store, WriteBatch, mem_store::MemStore},
         test_dag_builder::DagBuilder,
         transaction::TransactionConsumer,
